clean up , removed unused functions and interfaces
