+++ package.json
@@ -31,13 +31,13 @@
     "json-stable-stringify": "^1.0.2",
     "keccak": "^3.0.2",
     "loglevel": "^1.8.1",
-    "web3-utils": "^1.8.1"
+    "web3-utils": "^1.8.1",
+    "@toruslabs/fnd-base": "file:./toruslabs-fnd-base-v8.1.1.tgz"
   },
   "devDependencies": {
     "@babel/register": "^7.18.9",
     "@babel/runtime": "^7.20.1",
     "@toruslabs/eslint-config-typescript": "^1.2.0",
-    "@toruslabs/fnd-base": "file:./toruslabs-fnd-base-v8.1.1.tgz",
     "@toruslabs/fetch-node-details": "file:./toruslabs-fetch-node-details-v8.1.1.tgz",
     "@toruslabs/torus-scripts": "^1.3.0",
     "@types/chai": "^4.3.4",
