+++ src/Point.ts
@@ -16,21 +16,16 @@ class Point {
     this.ecCurve = ecCurve;
   }
 
-  // complies with EC and elliptic pub key types
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  encode(enc: string, params?: any): Buffer {
+  encode(enc: string): Buffer {
     switch (enc) {
       case "arr":
         return Buffer.concat([Buffer.from("0x04", "hex"), Buffer.from(this.x.toString("hex"), "hex"), Buffer.from(this.y.toString("hex"), "hex")]);
       case "elliptic-compressed": {
-        // TODO: WHAT IS THIS.?
-        let ec = params;
-        ec = this.ecCurve;
-        const key = ec.keyFromPublic({ x: this.x.toString("hex"), y: this.y.toString("hex") }, "hex");
+        const key = this.ecCurve.keyFromPublic({ x: this.x.toString("hex"), y: this.y.toString("hex") }, "hex");
         return Buffer.from(key.getPublic(true, "hex"));
       }
       default:
-        throw new Error("encoding doesnt exist in Point");
+        throw new Error("encoding doesn't exist in Point");
     }
   }
 }
