+++ src/constants.ts
@@ -1,0 +1,6 @@
+export const JRPC_METHODS = {
+  GET_OR_SET_KEY: "GetPubKeyOrKeyAssign",
+  COMMITMENT_REQUEST: "CommitmentRequest",
+  IMPORT_SHARE: "ImportShare",
+  GET_SHARE_OR_KEY_ASSIGN: "GetShareOrKeyAssign",
+};
