+++ src/helpers/errorUtils.ts
@@ -1,0 +1,1 @@
+export class GetOrSetNonceError extends Error {}
