+++ src/helpers/keyUtils.ts
@@ -1,0 +1,33 @@
+import BN from "bn.js";
+import { ec } from "elliptic";
+import createKeccakHash from "keccak";
+import { toChecksumAddress } from "web3-utils";
+
+import log from "../loglevel";
+
+export function keccak256(a: string | Buffer): string {
+  const hash = createKeccakHash("keccak256").update(a).digest().toString("hex");
+  return `0x${hash}`;
+}
+
+export function generateAddressFromPrivKey(ecCurve: ec, privateKey: BN): string {
+  const key = ecCurve.keyFromPrivate(privateKey.toString("hex", 64), "hex");
+  const publicKey = key.getPublic().encode("hex", false).slice(2);
+  log.info(publicKey, "public key");
+  const ethAddressLower = `0x${keccak256(Buffer.from(publicKey, "hex")).slice(64 - 38)}`;
+  return toChecksumAddress(ethAddressLower);
+}
+
+export function generateAddressFromPubKey(ecCurve: ec, publicKeyX: BN, publicKeyY: BN): string {
+  const key = ecCurve.keyFromPublic({ x: publicKeyX.toString("hex", 64), y: publicKeyY.toString("hex", 64) });
+  const publicKey = key.getPublic().encode("hex", false).slice(2);
+  log.info(key.getPublic().encode("hex", false), "public key");
+  const ethAddressLower = `0x${keccak256(Buffer.from(publicKey, "hex")).slice(64 - 38)}`;
+  return toChecksumAddress(ethAddressLower);
+}
+
+export function getPostboxKeyFrom1OutOf1(ecCurve: ec, privKey: string, nonce: string): string {
+  const privKeyBN = new BN(privKey, 16);
+  const nonceBN = new BN(nonce, 16);
+  return privKeyBN.sub(nonceBN).umod(ecCurve.curve.n).toString("hex");
+}
