+++ src/index.ts
@@ -1,3 +1,3 @@
+export * from "./helpers";
 export * from "./interfaces";
 export { default } from "./torus";
-export * from "./utils";
